t/line-filters/filtros-de-linea closes #31
