+++ public/escritura-de-archivos
@@ -279,7 +279,7 @@ hemos visto el uso de <code>stdin</code> y <code>stdout</code>.</p>
       
       
       <p class="next">
-        Siguiente ejemplo: <a href="line-filters">Line Filters</a>.
+        Siguiente ejemplo: <a href="filtros-de-linea">Filtros de línea</a>.
       </p>
       
       <p class="footer">
