+++ public/index.html
@@ -149,7 +149,7 @@
       
         <li><a href="escritura-de-archivos">Escritura de archivos</a></li>
       
-        <li><a href="line-filters">Line Filters</a></li>
+        <li><a href="filtros-de-linea">Filtros de línea</a></li>
       
         <li><a href="command-line-arguments">Command-Line Arguments</a></li>
